314 Refactor create barcode page
