+++ integration_tests/integration/createBarcode.spec.ts
@@ -1,7 +1,6 @@
 import IndexPage from '../pages/index'
 import AuthSignInPage from '../pages/authSignIn'
 import Page from '../pages/page'
-import CreateBarcodePage from '../pages/createBarcode'
 
 context('Create Barcode', () => {
   beforeEach(() => {
@@ -15,32 +14,9 @@ context('Create Barcode', () => {
     Page.verifyOnPage(AuthSignInPage)
   })
 
-  it('Create barcode tile exists', () => {
-    cy.signIn()
-    const indexPage = Page.verifyOnPage(IndexPage)
-    indexPage.createBarcodeLink().should('exist')
-  })
-
   it('Can enter create barcode page', () => {
     cy.signIn()
     const indexPage = Page.verifyOnPage(IndexPage)
-    indexPage.createBarcodeLink().click()
-    Page.verifyOnPage(CreateBarcodePage)
-  })
-
-  it('Create barcode page has prisoner input', () => {
-    cy.signIn()
-    const indexPage = Page.verifyOnPage(IndexPage)
-    indexPage.createBarcodeLink().click()
-    const createBarcodePage = Page.verifyOnPage(CreateBarcodePage)
-    createBarcodePage.prisoner().should('exist')
-  })
-
-  it('Create barcode page has continue button', () => {
-    cy.signIn()
-    const indexPage = Page.verifyOnPage(IndexPage)
-    indexPage.createBarcodeLink().click()
-    const createBarcodePage = Page.verifyOnPage(CreateBarcodePage)
-    createBarcodePage.continueButton().should('exist')
+    indexPage.clickCreateBarcodeLink()
   })
 })
